requirements.txt
python-dotenv
scikit-learn
numpy
opencv-python
packaging
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
torch
rfdetr
tensorboard
wandb
deep_sort_realtime
ipywidgets